src/test_new/requirements.txt
crewai>=0.28.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.0.0
